refactor: split MenuBarSection and MenuBarItem along their testable seams
Takes measured coverage from 74.10% to 81.09%, clearing the 80% statement
coverage criterion for the OpenSSF Best Practices silver badge.

Both files buried pure logic behind system-bound code, so excluding either
whole would have contradicted the rule stated at the top of
sonar-project.properties -- pure value types and algorithms stay measured even
under a UI-sounding path. Instead they are split the way MenuBarItemManager
already was when LayoutSolver, PendingLedger, AXIdentityCatalog and
ClickReactionVerifier were extracted out of it.

  MenuBarSectionLayout.swift (measured) takes Name, PresentationMode, notchGap
  and the three rules that decide how a section is shown. MenuBarSection.swift
  keeps show/hide/toggle, updateControlItemState and the rehide task and
  monitor, which need a live ControlItem/NSStatusItem, an AppState and a real
  NSScreen.

  MenuBarItem.swift (measured) keeps the value type. MenuBarItem+Enumeration
  .swift takes the window enumeration, which reaches the window server through
  Bridging/CGS and resolves owning processes over XPC.

No access level was widened. The unchecked initializers, the MenuBarItemTag and
Namespace helpers and ResolvedPID moved into the excluded file with the code
that calls them, so they stay private. `nonisolated` had to be restored on the
two nested enums -- an extension's nonisolated does not reach into nested
types. MenuBarItem's explicit memberwise init is now synthesized: it was only
written out because the unchecked initializers sat in the struct body and
suppressed synthesis.

Also excludes three files whose substance genuinely cannot run in CI:
AXItemActivator (Accessibility permissions; its two pure helpers were already
extracted and stay covered), SettingsDetailStyle (a PrimitiveButtonStyle whose
makeBody needs a ButtonStyleConfiguration, which has no public initializer),
and OnboardingTour (a SwiftUI View whose navigation lives in @State). The last
of those was contributing unstable coverage -- its lines were being counted
only when the test host happened to render the onboarding sheet at launch.

1889 tests pass, swiftlint --strict clean.
