+++ sonar-project.properties
@@ -35,6 +35,25 @@ sonar.exclusions = \
 # all of which stay measured and carry dedicated suites in ThawTests. New
 # decision logic belongs in those files, not back in the manager.
 
+# MenuBarSection.swift and MenuBarItem.swift were split the same way rather
+# than excluded whole, because both buried pure logic that the rule above says
+# must stay measured.
+#
+#   MenuBarSectionLayout.swift keeps the section names, the presentation modes,
+#   the notch gap and the three rules that decide how a section is shown.
+#   MenuBarSection.swift keeps the instance half -- show, hide, toggle,
+#   updateControlItemState, the rehide task and the mouse-moved monitor -- which
+#   needs a live ControlItem/NSStatusItem, an AppState and a real NSScreen.
+#
+#   MenuBarItem.swift keeps the value type: capability flags, auto-detected
+#   naming, identity, Equatable and Hashable. MenuBarItem+Enumeration.swift
+#   takes the window enumeration, which reaches the window server through
+#   Bridging/CGS and resolves owning processes over XPC. The unchecked
+#   initializers and the MenuBarItemTag/Namespace helpers moved with it rather
+#   than being widened to internal -- they are reachable only from that path.
+#
+# New decision logic on either type belongs in the measured file.
+
 # Entry points and app lifecycle
 sonar.coverage.exclusions = \
   Thaw/Main/AppDelegate.swift,\
@@ -60,7 +79,10 @@ sonar.coverage.exclusions = \
   Thaw/Utilities/SystemAppearance.swift,\
 \
   Thaw/MenuBar/MenuBarManager.swift,\
+  Thaw/MenuBar/MenuBarSection.swift,\
   Thaw/MenuBar/MenuBarItems/MenuBarItemManager.swift,\
+  Thaw/MenuBar/MenuBarItems/MenuBarItem+Enumeration.swift,\
+  Thaw/MenuBar/MenuBarItems/AXItemActivator.swift,\
   Thaw/MenuBar/Spacing/MenuBarItemSpacingManager.swift,\
   Thaw/MenuBar/MenuBarItems/MenuBarItemImageCache.swift,\
   Thaw/MenuBar/MenuBarItems/MenuBarItemServiceConnection.swift,\
@@ -78,11 +100,13 @@ sonar.coverage.exclusions = \
   Thaw/Permissions/PermissionsWindow.swift,\
   Thaw/Settings/Models/AppSettings.swift,\
   Thaw/Settings/SettingsPanes/*.swift,\
+  Thaw/Settings/SettingsDetailStyle.swift,\
   Thaw/Settings/SettingsView.swift,\
   Thaw/Settings/SettingsWindow.swift,\
   Thaw/Settings/UpdateConsentSheet.swift,\
   Thaw/Onboarding/OnboardingSheet.swift,\
   Thaw/Onboarding/OnboardingPermissionsView.swift,\
+  Thaw/Onboarding/Tour/OnboardingTour.swift,\
   Thaw/Onboarding/Tour/TourMockups.swift,\
   Thaw/Onboarding/Tour/WelcomeMockup.swift,\
   Thaw/Settings/Search/SearchViews.swift,\
